backend/db/requirements.txt
supabase==2.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
google-api-python-client==2.108.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pyjwt==2.9.0
yt-dlp==2024.8.6
openai==1.40.0
